Move AnsibleLintRule inside rules module
Moved code to ease reusability and avoid cyclic imports.
